Added Logo and "Pure" CSS Style to the Home Page
Added the Lauren's List logo to the home page.  Also added the pure
style css to the home page to give it an updated feel.  We can use this
style or find our own, but for now it is used in the example.  For more
information visit:  http://purecss.io/