Added Create Account Page
Added the create account page and changed some other inconsistencies
throughout the project.  I realize that we should direct the user to a
page where he selects if he is a contractor or a customer before he gets
to this page, but I just wanted to get this page out of the way first.
In other words, the links will most likely change.