Contractors Can Login and Logout! - Added Contractor.java and contractorCreateAccount.java
Also changed when customers create accounts, they are logged in
automatically.  Changed to check that all fields on customer form are
filled in. Added servlet info to web.xml.  Changes to form action for
contractorCreateAccount.html.  Changes to login.java, logout.java, and
homPage.java to accompany the additional user.